common: substitute "off_t" with "long long/size_t"
Modify libpmemlog and libpmemblk API to not use "off_t" type for
function arguments or returned value.

This modification is dictated by incompatible "off_t" definitions on
64-bit Linux and Windows operating systems.  On Windows, "off_t" is
always 32-bit long, which is not enough to hold offsets of large (>4GB)
pmem pool files.
To avoid potential problems with redefining "off_t" or with adding
additional typedefs, seems like the best choice is to use "long long"
or "size_t" types.

+++ src/include/libpmemblk.h
@@ -78,10 +78,10 @@ void pmemblk_close(PMEMblkpool *pbp);
 int pmemblk_check(const char *path, size_t bsize);
 size_t pmemblk_bsize(PMEMblkpool *pbp);
 size_t pmemblk_nblock(PMEMblkpool *pbp);
-int pmemblk_read(PMEMblkpool *pbp, void *buf, off_t blockno);
-int pmemblk_write(PMEMblkpool *pbp, const void *buf, off_t blockno);
-int pmemblk_set_zero(PMEMblkpool *pbp, off_t blockno);
-int pmemblk_set_error(PMEMblkpool *pbp, off_t blockno);
+int pmemblk_read(PMEMblkpool *pbp, void *buf, size_t blockno);
+int pmemblk_write(PMEMblkpool *pbp, const void *buf, size_t blockno);
+int pmemblk_set_zero(PMEMblkpool *pbp, size_t blockno);
+int pmemblk_set_error(PMEMblkpool *pbp, size_t blockno);
 
 /*
  * Passing NULL to pmemblk_set_funcs() tells libpmemblk to continue to use the
